+++ src/main/java/com/lowdragmc/mbd2/common/machine/definition/config/ConfigMachineSettings.java
@@ -5,14 +5,11 @@ import com.lowdragmc.lowdraglib.gui.editor.annotation.NumberRange;
 import com.lowdragmc.lowdraglib.gui.editor.configurator.IConfigurable;
 import com.lowdragmc.lowdraglib.syncdata.IPersistedSerializable;
 import com.lowdragmc.mbd2.common.trait.TraitDefinition;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.Singular;
+import lombok.*;
 import lombok.experimental.Accessors;
+import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
-import net.minecraft.world.item.ItemStack;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -20,6 +17,38 @@ import java.util.List;
 @Accessors(fluent = true)
 @Builder
 public class ConfigMachineSettings implements IPersistedSerializable, IConfigurable {
+    @Getter @Setter
+    public static class SignalConnection {
+        @Configurable(name = "config.machine_settings.signal_connection.front")
+        private boolean frontConnection = false;
+        @Configurable(name = "config.machine_settings.signal_connection.back")
+        private boolean backConnection = false;
+        @Configurable(name = "config.machine_settings.signal_connection.left")
+        private boolean leftConnection = false;
+        @Configurable(name = "config.machine_settings.signal_connection.right")
+        private boolean rightConnection = false;
+        @Configurable(name = "config.machine_settings.signal_connection.top")
+        private boolean topConnection = false;
+        @Configurable(name = "config.machine_settings.signal_connection.bottom")
+        private boolean bottomConnection = false;
+
+        public boolean getConnection(Direction front, Direction side) {
+            if (side == Direction.UP) {
+                return topConnection;
+            } else if (side == Direction.DOWN) {
+                return bottomConnection;
+            } else if (side == front) {
+                return frontConnection;
+            } else if (side == front.getOpposite()) {
+                return backConnection;
+            } else if (side == front.getClockWise()) {
+                return rightConnection;
+            } else if (side == front.getCounterClockWise()) {
+                return leftConnection;
+            }
+            return false;
+        }
+    }
     @Getter
     @Builder.Default
     @Configurable(name = "config.machine_settings.machine_level", tips = "config.machine_settings.machine_level.tooltip")
@@ -37,6 +66,11 @@ public class ConfigMachineSettings implements IPersistedSerializable, IConfigura
             "config.machine_settings.drop_machine_item.tooltip.2",
     })
     private boolean dropMachineItem = true;
+    @Getter
+    @Builder.Default
+    @Configurable(name = "config.machine_settings.signal_connection", subConfigurable = true,
+            tips = {"config.machine_settings.signal_connection.tooltip.0", "config.machine_settings.signal_connection.tooltip.1"})
+    private final SignalConnection signalConnection = new SignalConnection();
     @Singular
     @NonNull
     @Getter