+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 * rework the `ICapabilityProviderTrait` to allow a trait to provide multiple capabilities
 * add spark attachable to the mana trait
 * add redstone signal condition
+* add redstone signal feature (signal/direct signal, analog signal)
+  * you can set output signal by the node graph and kjs